get single item price successful
