+++ src/components/Verify.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Verify = () => {
+  return (
+    <div>Verfiy</div>
+  )
+}
+
+export default Verify
