+++ src/components/index.js
@@ -1,2 +1,4 @@
 export { default as Card } from './Card'
 export { default as ClassesCard } from './ClassesCard'
+export { default as PaystackIntegration } from './PaystackIntegration'
+export { default as Verify } from './Verify'
